Moved some types into a model module
