+++ src/core/data.scala
@@ -17,7 +17,7 @@
 
 package fury.core
 
-import fury._, io._, strings._, ogdl._
+import fury._, io._, strings._, ogdl._, model._
 import Graph.{CompilerDiagnostic, DiagnosticMessage, OtherMessage}
 import exoskeleton._
 import gastronomy._
@@ -54,20 +54,6 @@ object ManifestEntry {
 
 case class ManifestEntry(key: String, value: String) { def pairString: String = str"$key=$value" }
 
-object Kind {
-  implicit val msgShow: MsgShow[Kind] = v => UserMsg { t => v.name }
-  implicit val stringShow: StringShow[Kind] = _.name
-  val all: List[Kind] = List(Library, Compiler, Plugin, Application, Benchmarks)
-  def unapply(str: String): Option[Kind] = all.find(_.name == str)
-}
-
-sealed abstract class Kind(val name: String, val needsExecution: Boolean)
-case object Library extends Kind("library", false)
-case object Compiler extends Kind("compiler", false)
-case object Plugin extends Kind("plugin", false)
-case object Application extends Kind("application", true)
-case object Benchmarks extends Kind("benchmarks", true)
-
 object Module {
   implicit val msgShow: MsgShow[Module]       = v => UserMsg(_.module(v.id.key))
   implicit val stringShow: StringShow[Module] = _.id.key
@@ -98,49 +84,6 @@ object Binary {
 case class Binary(binRepo: BinRepoId, group: String, artifact: String, version: String) {
   def spec = str"$group:$artifact:$version"
   def paths(io: Io): Future[List[Path]] = Coursier.fetch(io, this)
-}
-
-object Permission {
-  implicit val msgShow: MsgShow[Permission] = stringShow.show
-  
-  implicit val stringShow: StringShow[Permission] =
-    p => str"${p.className} ${p.target} ${p.action.getOrElse("-"): String}"
-  
-  implicit def diff: Diff[Permission] = Diff.gen[Permission]
-
-  val Classes: List[String] = List(
-    "java.awt.AWTPermission",
-    "java.io.FilePermission",
-    "java.io.SerializablePermission",
-    "java.lang.management.ManagementPermission",
-    "java.lang.reflect.ReflectPermission",
-    "java.lang.RuntimePermission",
-    "java.net.NetPermission",
-    "java.net.SocketPermission",
-    "java.net.URLPermission",
-    "java.nio.file.LinkPermission",
-    "java.security.AllPermission",
-    "java.security.SecurityPermission",
-    "java.security.UnresolvedPermission",
-    "java.sql.SQLPermission",
-    "java.util.logging.LoggingPermission",
-    "java.util.PropertyPermission",
-    "javax.management.MBeanPermission",
-    "javax.management.MBeanServerPermission",
-    "javax.management.MBeanTrustPermission",
-    "javax.management.remote.SubjectDelegationPermission",
-    "javax.net.ssl.SSLPermission",
-    "javax.security.auth.AuthPermission",
-    "javax.security.auth.kerberos.DelegationPermission",
-    "javax.security.auth.kerberos.ServicePermission",
-    "javax.security.auth.PrivateCredentialPermission",
-    "javax.sound.sampled.AudioPermission",
-    "javax.xml.bind.JAXBPermission",
-    "javax.xml.ws.WebServicePermission"
-  )
-}
-case class Permission(className: String, target: String, action: Option[String]) {
-  def hash: String = this.digest[Sha256].encoded[Hex].toLowerCase
 }
 
 object PermissionHash {
@@ -992,7 +935,7 @@ case class Hierarchy(schema: Schema, dir: Path, inherited: Set[Hierarchy]) {
       
       allProjects          <- conflictIds match {
                                 case x if x.isEmpty => Success(projects ++ nextProjects)
-                                case _ => Failure(ProjectConflict(conflictIds, h1 = this, h2 = hierarchy))
+                                case _ => Failure(ProjectConflict(conflictIds/*, h1 = this, h2 = hierarchy*/))
                               }
     } yield allProjects
 
@@ -1199,33 +1142,6 @@ object SchemaId {
 
 case class SchemaId(key: String) extends Key(msg"schema")
 
-object ProjectId {
-  implicit val msgShow: MsgShow[ProjectId]       = p => UserMsg(_.project(p.key))
-  implicit val stringShow: StringShow[ProjectId] = _.key
-  implicit def diff: Diff[ProjectId]             = (l, r) => Diff.stringDiff.diff(l.key, r.key)
-
-  def parse(name: String): Try[ProjectId] = name match {
-    case r"[a-z](-?[a-z0-9]+)*" => Success(ProjectId(name))
-    case _                      => Failure(InvalidValue(name))
-  }
-}
-
-case class ProjectId(key: String) extends Key(msg"project")
-
-object ModuleId {
-  implicit val msgShow: MsgShow[ModuleId]       = m => UserMsg(_.module(m.key))
-  implicit val stringShow: StringShow[ModuleId] = _.key
-  implicit def diff: Diff[ModuleId]             = (l, r) => Diff.stringDiff.diff(l.key, r.key)
-  final val Core: ModuleId                      = ModuleId("core")
-
-  def parse(name: String): Try[ModuleId] = name match {
-    case r"[a-z](-?[a-z0-9]+)*" => Success(ModuleId(name))
-    case _                      => Failure(InvalidValue(name))
-  }
-}
-
-case class ModuleId(key: String) extends Key(msg"module")
-
 object Repo {
   implicit val msgShow: MsgShow[Repo]       = r => UserMsg(_.url(r.simplified))
   implicit val stringShow: StringShow[Repo] = _.simplified
@@ -1320,14 +1236,6 @@ case class SourceRepo(id: RepoId, repo: Repo, track: RefSpec, commit: Commit, lo
   def sourceCandidates(io: Io, layout: Layout)(pred: String => Boolean): Try[Set[Source]] =
     listFiles(io, layout).map(_.filter { f => pred(f.filename) }.map { p =>
         ExternalSource(id, p.parent): Source }.to[Set])
-}
-
-case class BinRepoId(id: String)
-
-object BinRepoId {
-  implicit val msgShow: MsgShow[BinRepoId] = v => UserMsg(_.repo(v.id))
-  implicit val stringShow: StringShow[BinRepoId] = _.id
-  final val Central: BinRepoId = BinRepoId("central")
 }
 
 case class Repo(url: String) {
@@ -1636,5 +1544,3 @@ object Parameter {
 }
 
 case class Parameter(name: String) { def parameter = str"-$name" }
-
-abstract class Key(val kind: UserMsg) { def key: String }
